split into primitives and pallet
