+++ parachain/pallets/erc20-app/src/mock.rs
@@ -17,10 +17,8 @@ use sp_runtime::{
 	DispatchError, MultiSignature,
 };
 
-use snowbridge_core::{
-	assets::{RemoteParachain, XcmReserveTransfer},
-	ChannelId,
-};
+use snowbridge_core::ChannelId;
+use snowbridge_xcm_support_primitives::{RemoteParachain, XcmReserveTransfer};
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
