+++ parachain/pallets/eth-app/src/lib.rs
@@ -38,10 +38,8 @@ use sp_core::H160;
 use sp_runtime::traits::StaticLookup;
 use sp_std::prelude::*;
 
-use snowbridge_core::{
-	assets::{RemoteParachain, XcmReserveTransfer},
-	ChannelId, OutboundRouter,
-};
+use snowbridge_core::{ChannelId, OutboundRouter};
+use snowbridge_xcm_support_primitives::{RemoteParachain, XcmReserveTransfer};
 
 pub use pallet::*;
 use payload::OutboundPayload;
@@ -152,7 +150,7 @@ pub mod pallet {
 		) -> DispatchResult {
 			let who = T::CallOrigin::ensure_origin(origin.clone())?;
 			if who != <Address<T>>::get() {
-				return Err(DispatchError::BadOrigin.into());
+				return Err(DispatchError::BadOrigin.into())
 			}
 
 			let recipient = T::Lookup::lookup(recipient)?;
@@ -169,7 +167,7 @@ pub mod pallet {
 							destination.para_id,
 							err
 						);
-						return TransactionOutcome::Rollback(DispatchError::Other("foo").into());
+						return TransactionOutcome::Rollback(DispatchError::Other("foo").into())
 					}
 					TransactionOutcome::Commit(Ok(()))
 				});
