+++ parachain/pallets/xcm-support/Cargo.toml
@@ -1,13 +1,12 @@
 [package]
-name = "snowbridge-xcm-support-primitives"
+name = "snowbridge-xcm-support"
+description = "Snowbridge XCM Support Pallet"
 version = "0.1.1"
 authors = [ "Snowfork <contact@snowfork.com>" ]
 edition = "2021"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false }
-scale-info = { version = "2.0.1", default-features = false, features = [ "derive" ] }
-serde = { version = "1.0.136", optional = true, features = [ "derive" ] }
 
 frame-support = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.19", default-features = false }
 frame-system = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.19", default-features = false }
@@ -20,14 +19,12 @@ xcm = { git = "https://github.com/paritytech/polkadot.git", branch = "release-v0
 xcm-executor = { git = "https://github.com/paritytech/polkadot.git", branch = "release-v0.9.19", default-features = false }
 xcm-builder = { git = "https://github.com/paritytech/polkadot.git", branch = "release-v0.9.19", default-features = false }
 
-snowbridge-core = { path = "../../primitives/core", default-features = false }
+snowbridge-xcm-support-primitives = { path = "../../primitives/xcm-support", default-features = false }
 
 [features]
 default = [ "std" ]
 std = [
         "codec/std",
-        "scale-info/std",
-        "serde/std",
         "sp-std/std",
         "sp-runtime/std",
         "frame-support/std",
@@ -35,6 +32,6 @@ std = [
         "xcm/std",
         "xcm-executor/std",
         "xcm-builder/std",
-        "snowbridge-core/std",
+        "snowbridge-xcm-support-primitives/std",
         "pallet-xcm/std"
 ]
